Add CONTENT_HASH option for writing paths to buck-out
Summary: Right now, we always write build artifacts to paths with a configuration hash. Now, a path can be specified as wanting to be written to a path with its content-based hash, and in that case we do not include the configuration hash. Instead, we require the caller to specify the content-based hash and we write it _after_ the package, before the short name.

Reviewed By: JakobDegen

Differential Revision: D72249286

fbshipit-source-id: cb882fa42caa9f0b7775e7925f5a5665800d11db