Allow outputs to declare themselves to use content-based path hashing
Summary:
As discussed in https://fb.workplace.com/groups/buck2dev/posts/3975810866040289/, we tie whether or not we're using content-based path hashing to an output, rather than to an action.

I am using `uses_experimental_content_based_path_hashing` for now because I need people to know it is experimental (right now, if they turn it on it won't work correctly, but I need the option available for tests etc).

Reviewed By: JakobDegen

Differential Revision: D72457376

fbshipit-source-id: f5ac91310b02273791ff23a1dc8e6ae91571a52a