Support streaming output stdout in superconsole and simple console
Summary:
RFC: https://fb.workplace.com/groups/buck2dev/permalink/3976662752621767/

This  diff adds a `StreamingOutput` event and handles the event in superconsole and simple console.

It adds a `StreamingOutputWriter` in bxl and used in `ctx.output.stream` and in `command.rs`.
We use `StreamingOutputWriter` for all outputs in `command.rs` to avoid potential ordering issues that can occur when mixing output from `StreamingOutputWriter` and `PartialOutput`.

Reviewed By: cjhopman

Differential Revision: D73566861

fbshipit-source-id: 7f389a257d8c11e0cd6b6aea701155316e66e7b6