HACKING.md: document running checks and tests
This was missing info that is what external contributors would want
to find to be able to do core development tasks without reading the CI
definitions.

This also reenables clippy on macOS since it seems to work on my
machine, probably was fixed upstream.

CC: https://github.com/facebook/buck2/issues/948, which still needs a
better error message
CC: https://github.com/facebook/buck2/issues/943, which is still a
footgun, but at least it's documented how to do correctly now