Add content_hash function to ArtifactValue
Summary:
For directories and files, this is simple - we just use the same value as calculated by `digest()`.

For symlinks, we can use the `deps`, unless the symlink points to itself (I think that's the only case it doesn't have `deps`, but not 100% sure). We fallback to calculating a hash based on the symlink path.

For external symlinks, just return a hash of the path.

Reviewed By: JakobDegen

Differential Revision: D72457373

fbshipit-source-id: 9c703776a1f7b6106b00e052285b84eb0631e9bc