Sort dep_files content
Summary:
This is a follow-up on D74407171, which ensures that the dep_files context is always sorted before saving to a file, thereby preventing non-deterministic dep-files.

I suspect that KSIC may also contribute to non-deterministic dep-files, as recompiled files depend on changes made by the user. I find it easier to sort the requested output as all intermediate steps.

Differential Revision: D74407171

fbshipit-source-id: a864d61abfaba950631c0a534c819c82d1f26021