Change the TCP port direction in RPC client/server
Summary:
Previously, the HealthCheckClient was setting up the TCPListener and the CLI would connect to it despite the CLI being the health check server. Although semantically it is incorrect, this works somewhat OK considering that the grpc just cares about the stream and not necessarily about the direction of TCPListener/TCPStream. However, the python grpc server requires the port direction to be set as listener.

Fixing the TCP port direction in this diff based on comments in https://fburl.com/code/tqyrmy16

Reviewed By: blackm00n

Differential Revision: D74463944

fbshipit-source-id: 5ce9e1f54ca961c3bf1001ecae667a8b5018c58e