Simplify python_bootstrap_binary
Summary:
Instead of setting PYTHONPATH, we can just copy the main binary into the
right location relative to the libraries and run it from there, so that
we implicitly set the path correctly.

We can no longer have a binary and a library defining the same source
file, but that wouldn't work correctly previously anyway (the library
source file could never be imported since the binary would take
precedence). Fix that pattern in a few places.

Reviewed By: JakobDegen

Differential Revision: D70819181

fbshipit-source-id: ae44c4a4b30d63e9f21c1146f752936efb96c94c