Don't set preferred_linkage in go_exported_library.bzl
Summary: I spoke with podtserkovskiy, turns out `c_shared` is supposed to output a single shared library with its dependencies statically linked in. Setting `preferred_linkage` in `merged_link_infos` would propagate the link style to the library's dependencies as well, let's get rid of it and rely on the `LinkStyle` set in the actual link step instead.

Reviewed By: capickett

Differential Revision: D74486179

fbshipit-source-id: 5f94af8170f8c2eec51952cce4caf2da8c17bd3a