Do not include dependent artifacts with bytecode manifest
Summary: When we build inplace pars we need to make sure the manifest does not pull in any dependent artifacts or the action won't run until those artifacts are built and materialized. This allows bytecode compilation and inplace par building to run in parallel. This also brings the behavior of bytecode manifests in line with other manifests.

Reviewed By: IanChilds, AishwaryaSivaraman

Differential Revision: D74600773

fbshipit-source-id: 276fac1a8057fd283360295e7143fe0a4aac9761