Toggle default for is_deferrable to true
Summary: `is_deferrable` currently seems strictly better than the alternative and we intend to opt users into it accordingly (see task), but we are aware that there may be unforeseen side effects. Toggling the default for those not using the option is the minimal (easily revertable) change to test the waters before starting to remove the option.

Reviewed By: IanChilds, krallin

Differential Revision: D74501121

fbshipit-source-id: fb4f34c906409cd7f1246b1c72c677da3beb3cb3