Fix connection error
Summary:
As [reported](https://fb.workplace.com/groups/buck2users/permalink/3964801990442819/) by users, the health check client fails to connect to the health check server.

Problem:
When running a build, health check server creates a state info file with the listening TCP port, and exits after the build completes. When a new build starts, the health check client tries to connect directly to the port mentioned in the state file. If the new health check server (for the new command), does not write to the file before the client reads it, the client tries to connect to a previous non-existent TCP port and thus failing.

Fix:
Reset state information before builds.

Reviewed By: samkevich

Differential Revision: D74660490

fbshipit-source-id: b10ff4e88ef712d7ccb3d25cbc62bce4633a8bf9