Add content-based path support for download_file
Summary: If we have the metadata supplied, then use that to resolve the path correctly. In the case that we do not have metadata, we just say that content-based paths cannot be used (we could alternatively download the file to a temporary destination and then calculate the content hash before moving it, but the metadata case is the better one anyway so seems reasonable to make this restriction, especially for now).

Reviewed By: scottcao

Differential Revision: D74252141

fbshipit-source-id: cad7a03c26f6da4c515136379e5bdad108506714