Fix handling of adb-executable-path
Summary:
The diff [D74108825](/intern/diff/74108825/) added the `adb-executable-path` option to the installer's command line options, which helped avoid configuration changes due to `-c android.adb`.
Although it fixed the precedence of locations to loo for `adb`, if `adb` isn't found in any of the default locations, like PATH or `/opt/android_sdk/platform-tools/adb`, it would result in an error.
This issue led to failures in Sandcastle jobs, such as [this one](https://www.internalfb.com/sandcastle/workflow/63050394803005871). This diff addresses and fixes the problem.

Reviewed By: christolliday

Differential Revision: D74825329

fbshipit-source-id: 9b0adcc12028711bd0838a4d8b9f578d3de7e5c8