move builder common code in a separate struct
Summary:
All the `GenericREClientBuilder` does is independent from the factory trait.

Create a common struct and a common trait with default implementation and refactor the current builder

NOTE: due to Rust dynamic dispatching we now need to `use` everywhere the new trait `ClientBuilderCommonMethods`. This means changing the imports of all the callers. I investigated if there's a better way to do this and maybe one way could be to have a `prelude` module and let clients `use prelude::*`.
I think it's not necessary to do this in this diff though since we already let clients import a bunch of different things

Reviewed By: c-ryan747

Differential Revision: D74482810

fbshipit-source-id: d414ee7d46de059e5b7312a9c7310d1c498463b4