Add content-based path support for local actions
Summary:
This adds basic support for running local actions with content-based paths (either as inputs or outputs). For the inputs, we just calculate their path correctly based on the `ArtifactValue` that we have for them. For outputs, we use a placeholder path when running the action, and then we copy the output to the correct path (if necessary) once the action has been run.

There's still a case where we can be using the same output path while two local actions are running at the same time (since the placeholder path is the same regardless of configuration). I'll fix that up in a future diff.

Reviewed By: cjhopman

Differential Revision: D74252137

fbshipit-source-id: d57f3a29473d32ff7d24a589b7b41a0f58803e52