Refactor local resource calculation such that we resolve the command line after computing inputs
Summary: With content-based paths, we cannot resolve a command line until we have built all the artifacts that are used on the command line. So, instead of visiting artifacts and resolving the command line in `required_local_resources_setup_contexts`, we move all of that into `prepare_local_resource`, which is where we already build the inputs.

Reviewed By: samkevich

Differential Revision: D74577945

fbshipit-source-id: f5bda3c9efa9b85b1ce564d4b72461736bbcfaa6