Split expander::expand in two
Summary:
With content-based paths, we can't resolve a command line without having all the inputs to that command line (otherwise we don't know the content-hash to use when resolving that paths of those inputs).

Therefore, in preparation for supporting content-based paths, we split `expand` into 2 parts - first we get the inputs, then we expand the command lines etc.

Reviewed By: blackm00n

Differential Revision: D74468266

fbshipit-source-id: c83b9bbb5b3cca2e82eb9a25a0cc0d6ceaa36d8d