let starlark evaluations during analysis be canceled
Summary:
Today, starlark evaluations must run to completion. This inserts an "infrequent instruction check" in the evaluator that is called every so often, in which we now check whether the `CancellationContext` has been notified. If so, we return with `EvaluatorError:Cancelled`.

#internal
This should allow eagerbuild's use of the `--premptible` flag to exit early from starlark evaluations and not hold up the next build in "synchronization and waiting".

This also provides a good spot for T219887296 to be addressed

Reviewed By: JakobDegen

Differential Revision: D73694716

fbshipit-source-id: 718253cf9e37fc8fee3de60f9f3075211f534b19