expose cancellation requests from CancellationContext
Summary:
Cancellation `Context/Observer/Handle/Future` normally have two separate shared data parts, and only the `Future` can see both. Normally the Handle is used to signal that the future should be cancelled (via the bool in `SharedStateData`), the `poll()` method on the `Future` will see that cancellation "request" and propagate the cancellation into the other shared `notified: AtomicU8`, and wake any waiting `CancellationContext/CancellationObserver`s so that the async task can be woken, detect cancellation and return.

For a synchronous computation like a StarlarkEvaluation, we never yield back up to the task's Future, so it's `poll()` method never gets a chance to propagate the cancellation state.  So this adds `Arc<SharedStateData>` pointers into the data structures visible to the `CancellationContext/CancellationObserver` objects, which the starlark evaluator can poll to determine if it should exit early with an error.

Reviewed By: cjhopman

Differential Revision: D74433525

fbshipit-source-id: d353740f2d744fdf3e06343493175b34dd2fe57c