Ban get_artifact_path_display with content-based paths
Summary:
The way that BXL is written right now is that ensured artifacts generally do not have access to their content hash, and so we are not able to correctly resolve content-based paths.

For now, we just error if we try to do so. In future, we'll have to refactor this to make it work correctly if we want to roll out content-based paths in more places.

Reviewed By: blackm00n

Differential Revision: D74808720

fbshipit-source-id: c8bca435a68e100cb5f6dda83266becdcc099998