Add some install support for content-based paths
Summary:
Right now, content-based paths cannot be supported for `install`, because we pass the paths before they have been built in our initial message to the installer.

I don't think it's necessary to pass the paths at that point, so we could refactor this later on to allow that, but for now let's just ban using content-based paths in installation.

I did fill in the other parts so that they will work once we allow content-based paths.

Reviewed By: cjhopman

Differential Revision: D74470306

fbshipit-source-id: b445675c5142b89d0ee186b42592811ec2ddc9cf