Make sure that we don't use ignore_artifacts with content-based paths
Summary: It doesn't make sense to use `ignore_artifacts` with content-based paths, since that would mean that we are unable to resolve the paths to the artifacts. Therefore, if `ignore_artifacts=True` then we create a visitor that just verifies that none of the artifacts that are being ignored have content-based paths.

Reviewed By: cjhopman

Differential Revision: D74573540

fbshipit-source-id: d5751c25631f626bb5225a32213f8a8dcfe879c8