Add rust type for the aggregated metrics
Summary:
To pass the metrics to build report, we need the buck2_core types but it is presently lost during protobuf conversion e.g. The ConfiguredTargetLabel used in build report is not easy to reconstruct from the buck2_data::TargetLabel.

Creating new rust equivalent types for the protobuf types in this diff. Usage in further diffs.

Reviewed By: blackm00n

Differential Revision: D74882335

fbshipit-source-id: 785bfecdc501456c1695ab78cb5e8f67028585a7