Don't allow local actions with the same output paths to execute at the same time
Summary:
When using content-based paths, the temporary output paths are not unique across different configurations. That means that when we run locally, there is a risk of using the same path in multiple different actions.

To avoid that, we make sure that we only ever run one action per content-based output path and content-based scratch path by setting them in the host_sharing_requirements, using the `OnePerTokens` functionality added in D74479880.

Reviewed By: cjhopman

Differential Revision: D74480906

fbshipit-source-id: bb0af0db0a51768a4a2c01c031100d3789952fd8