Add content-based path support for relative_to
Summary: If the artifact that we want to calculate our path relative to can be calculated with its content-hash, then we use that. Otherwise, we rely on the construction of the content based path always being of the same form (i.e. `package/<content_hash>/short_name`) and we just use a placeholder for the content hash, which results in a correct relative path.

Reviewed By: blackm00n

Differential Revision: D74480914

fbshipit-source-id: 348258e87d98af9b6cba1e6543b4c36d2480a1d4