Use different scratch path for content-based outputs
Summary: The scratch path currently contains the (hashed) configuration as part of its path. The idea of content-based paths is that ultimately we want to be able to de-dupe actions, and we won't be able to do that if the scratch path contains the configuration. So, if all the outputs are content-based, then we make the scratch path content-based too (if they are not all content-based then the action will contain the configuration already so it doesn't matter if the scratch path contains it too).

Reviewed By: cjhopman

Differential Revision: D74463940

fbshipit-source-id: 05fd058e365f1d03ba9c4e145f3b91cf97c354a4