Further clarify cert error message on Windows
Summary:
As title, we've gotten quite a few reports on Windows Users group (see [post](https://fb.workplace.com/groups/winbook/permalink/3907943359478714/)) from this message.
Based on the comments users are running the `getchef` command on a non-admin terminal which doesn't fix their issue.
Here I'm modifying the error to explicitly mention they should run this on an admin PowerShell.

I know this seems pretty hacky, but wanted to start a discussion on whether this is good and get suggestions if this is plain bad :D

Reviewed By: JakobDegen

Differential Revision: D75090979

fbshipit-source-id: b20605b7a891fb63ac33253731ab42b088796f82