Handle universal archives in dist_lto_prepare
Summary:
Builds may check in universal archives. Distributed thin-lto links cannot operate directly on static archives, they need to be unarchived first. Distributed thin-lto uses the ar tool associated with the cxx_toolchain to operate on the archive, but it can't properly do this for universal archives, those need to be manipulated with lipo.

This diff modifies the "thin_lto_prepare" action to check if an archive is universal, and if so extract the slice for the target architecture first, then proceed as usual.

Reviewed By: drodriguez

Differential Revision: D75039342

fbshipit-source-id: 0edd00e7c20320730597c8e5ca65350901881d87