Attempt at keep going
Summary:
I want to add a flag to the arc pyre invocation that allows it to keep going when it encounters a target that is broken or unable to be built. One area where it would be really beneficial would be in our upgrade infra, where one broken target can stop a run for thousands of others.

Would love initial feedback on an approach like this.

Reviewed By: ndmitchell

Differential Revision: D75101064

fbshipit-source-id: 0ddb7e46f34233bb98c2a9f2a29b51519e080708