Make dep_files tests check MatchDepFiles events when falling back to running locally
Summary: Added a comment to make it clear why we check that we _either_ got a remote dep file hit or we ran locally. Also, made it so that when we run locally we check that we just ran a `MatchDepFiles` for the identical lookup, and we didn't get a remote dep file cache hit but failed the verification.

Reviewed By: blackm00n

Differential Revision: D75525431

fbshipit-source-id: 38eb9e326e8a12d80bf21d56e80e0aef49879cab