Fix for a resource broker setup failure
Summary:
I've seen the resource broker fail when running iOS tests locally (ex. P1825551980). The failure is caused by `simctl` reporting a simulator in a state (ex. 'Creating') that isn't recognized by the resource broker.

I looked inside the `CoreSimulator` framework binary and found all of the possible states that it can report. This diff adds the missing states to the states recognized by the resource broker so that it won't die if a simulator is in one of those states when the simulator list is generated.

Reviewed By: cpepin

Differential Revision: D75544568

fbshipit-source-id: d02925b5489ac14967310c5ea8eb1a92eddb5d72