Add content-based support for dep-file command lines
Summary:
When we do a lookup for a dep-file hit, we check that the command line is the same. This is a problem because content-based paths change the command line, but it doesn't necessarily mean that the dep-file is not valid (e.g. if the command was `[<binary>, <output>, <unused_input_path>]`).

This diff fixes that by constructing a "dep-file command line" that uses a placeholder instead of a content-hash. It's a bit inefficient, other ideas are welcomed.

Reviewed By: blackm00n

Differential Revision: D74584268

fbshipit-source-id: ab2839e6cc7bf177895912975d4fac0521cae0ce