Add SetSketch to GraphSketch struct
Summary:
Context: This stack enables the logging of the total configured graph size for the entire build to the build report.

To achieve this, we take the approach of merging all the individual graph sketches of each top level configured target into one sketch. There are a couple reasons we take this approach rather computing a fresh sketch by traversing the deps of all configured target nodes.
1. There is no place in the build code where we get all top level targets as `ConfiguredTargetNode`s.
2. On CI, we already have to compute a configured sketch for every top level node, so this cost is already paid for. Given sketches have bound sizes, merging sketches of all top level targets shouldn't be much more expensive than computing sketch for the whole graph in the worst case.

Since `SetSketch` only supports merge with other `SetSketch`es, we need to add the `SetSketch` to the return type of the dice call for computing configured graph sketch. This diff adds `SetSketch` to `GraphSketch` struct and renames it `MergeableGraphSketch`. `MergeableGraphSketch` holds both the `SetSketch` and a precomputed signature in bytes for implementing `Hash` and `Eq`. D75418406 adds the ability for a `VersionedSketcher` to merge a `MergeableGraphSketch`.

Reviewed By: IanChilds

Differential Revision: D75418407

fbshipit-source-id: 97d4e5117bd3f45a64d9703ebed8ebe509084080