Write a intro "What is buck2" to the start of buck2 101 series
Summary:
Write a intro "What is buck2" to the start of buck2 101 series

It just cover very basic thing and not introduce too many details about buck2.

There is some overlapping between https://buck2.build/docs/about/why.
https://buck2.build/docs/about/why has a problem that is have too many texts

My plan is to make this to be the landing page of this content, but need more time to design this.

Reviewed By: IanChilds

Differential Revision: D74702868

fbshipit-source-id: b3718ba9cd2c232cba2650f4efc63e4fbf22a149