Typo suggestion for DirectoryDoesNotExist error
Summary:
Looking at Time to Fix data, DirectoryDoesNotExist is one of the most common errors with a relative high time to fix considering that it should be a very simple fix.

Examining some of these errors myself, it looks like the most are caused by either typos or wrong directory/cell. Using levenshtein distance to help with typos, dealing with wrong directory/cell in the next diff

Reviewed By: JakobDegen

Differential Revision: D74903042

fbshipit-source-id: 6b57ee1ce9472d69f84e0d362b4946d39694c4d9