Add uses_experimental_content_based_path_hashing = true to all action declarations that create output artifacts
Summary:
It's kinda tedious to create an artifact and then pass it into the action creation, just so that you can turn on `uses_experimental_content_based_path_hashing = True`. Instead, just add it as a parameter on actions that create output artifacts, and use it when creating them.

We do verify that if the output was already created and you set it on the action too, that what you set on the action is consistent with the output artifact.

Reviewed By: JakobDegen

Differential Revision: D75676939

fbshipit-source-id: 04f0d8b182cc9aeb64f071b2bac2f9324700bfa5