fix concat jars unit test
Summary:
Tests were failing for a while.
Concatenated jars can have multiple entries with same name, the dedup hapens at runtime, when the ClassLoader trnsform the entries into an index using a Map.

So the test should not be fixed into the order and in the fact that all entries exist and the content is of the expected value.

Reviewed By: RuslanLa

Differential Revision: D75653895

fbshipit-source-id: b2e9bc5994d9c2f884282894f0ff75fa3749c835