bxl: build just once
Summary:
Keep `BuildResult` and result of ensure around and avoid calling the APIs multiple times.

The old code also forgot to include `suite` attribute in the set of things that will be prebuilt - without including this, some builds would fail.

This seems to cut down incremental execution by ~200ms or 20%.

Reviewed By: TheGeorge

Differential Revision: D75869434

fbshipit-source-id: 9e7239e4796683982bbe287d80341aa514b3ab46