buck2_daemon: automatically max out fd limits on startup (#986)
Summary:
This automatically increases the fd limit of the buck2d process up to whatever the existing user hard limit is, or a reasonable limit (0x80000) if the soft limit is already good enough. This is not configurable and it always happens without question.

The logic behind this is simple: there is never, ever, ever a scenario where I am thinking to myself: "You know what would be great? If my build system ran out of file descriptors and the build failed". Yet this happens all the time with on buck2 on macOS and (apparently?) my Fedora Asahi laptop, where the default soft limits are very, very low.

When I run `buck2 build`, I am asking it to build things. Buck is good at building things, and often it builds lots of things. The soft limit is a simple guardrail, but only that. And if too many files are open due to something like an excessive number of actions, the reality is that there is very little chance the file descriptor table is actually the limiting resource, versus the underlying machine resources (net/disk/cpu/memory).

Just always set the rlimit to the hard cap. This will be inherited by children including the forkserver, which is completely fine.

Fixes https://github.com/facebook/buck2/issues/419. Closes https://github.com/facebook/buck2/issues/928. Closes https://github.com/facebook/buck2/issues/981.

Pull Request resolved: https://github.com/facebook/buck2/pull/986

Reviewed By: iguridi

Differential Revision: D75642106

Pulled By: JakobDegen

fbshipit-source-id: bcd4203fbbf11f4841e49535093fa7209b724d7b