Use ContentBasedPathHash::OutputArtifact in resolve_outputs
Summary:
This function is only used when uploading locally run actions to the action cache: https://www.internalfb.com/code/fbsource/[85ba5e4edfe6edd3a2bfb73a10f735aeeed511ff]/fbcode/buck2/app/buck2_execute_impl/src/executors/caching.rs?lines=319

The important thing here is that the action definition is for outputs that used a placeholder for the content-based hash (since it is not known ahead of time). So, in order for this to work correctly, we need to use those output paths when uploading outputs.

Reviewed By: samkevich

Differential Revision: D75866194

fbshipit-source-id: 270e58b543e9dcb17943184f992f40cec3e81b83