repro an eden bug
Summary:
If a user calls `hg up -C` eden doesn't count all cleaned files as changes, although it does for `hg revert --all`. It's a bug in eden due to how eden is handling removing the file internally when you call update. And it should be fixed in D75817781
Because rollout of this fix can take a while I want to have a test in buck2 that will repro the bug to see when it will be fixed.

Reviewed By: IanChilds

Differential Revision: D75885402

fbshipit-source-id: dafb51f76061092082ea89161aba2c8a2ff4ad35