toolchain: Move interpreter manifest entry to toolchain
Summary:
And out of the macro

In principle, it's 1000x easier to get this value correct here than it is in the macro, because it's just lying around already.

Note that this is not 100% behavior preserving - for example, it doesn't have the completely incorrect windows host_info check

Unfortunately, in order to make this landable, we have to introduce a hack into the prelude: The problem is that outside of the [fbcode macros](https://fburl.com/code/ztnogq2j), we don't use any of the starlark side stuff to generate the manifest, but instead rely on the par builder generating it for us.

The right way to fix this hack is to move the entirety of the manifest generation stuff either out of the macros completely and into the toolchain, or at least into a shared macro layer where it also applies to the rest of fbsource. Unfortunately, doing so is not super easy while being sure not to break anything, particularly because there's currently two par builders floating around. Might be a nice bit of better engineering to pick up after that (and will bring load time wins if it's moved to the toolchain)

Reviewed By: zsol

Differential Revision: D75915712

fbshipit-source-id: df8e1e4e8e5fb97849d1bc33a618ff4fe7f630f6