Don't implement std Error for BxlEvalError
Summary:
The implementation here means that BxlEvalError implements
Into<buck2_error::Error>, but the implementation of that actually loses like
all the error information (because it just gets the Display from BxlEvalError
and nothing from the inner buck2 error).

This is a really subtle footgun. Let's remove the impl. This also makes the
existing conversions actually preserve the full buck2_error.

Reviewed By: Will-MingLun-Li

Differential Revision: D75909735

fbshipit-source-id: 5f4b6432bf2923f66a16e97beb942497703abab3