screenbrightness library
Summary:
# "Disable" monitor on Mac

The private API to disable Mac's monitor is not stable and might not work in some cases. We decided to disable Mac's monitors by creating a virtual monitor, and let all physical monitors mirror it. In this scenario, we will not be blocked by physical monitor's resolution list and we can achieve ultrawide monitor with fully virtualized monitors setup. After mirroring, we will need to blackout the Mac screens. For Mac built-in screen, there is a private API to do so. For external monitors, we need to communicate with the monitor via Display Data Channel (DDC) to adjust the brightness. DDC can set the brightness of monitor to 0 and this will black out all LED monitors, not LCD, so we need to set the gamma table of the monitor and change the color profile to 0 to finalize the last piece of blackout.

# Summary
Creates a new module to:
1. Set the screen physical brightness via DDC.
2. Set software brightness via gamma table.
3. Set the Mac built-in screens brightness via the private API `DisplayServicesSetBrightness`.

This library refers to https://github.com/MonitorControl/MonitorControl. This diff is not a direct copy/paste the github project but I refer to the functions it is using.

Reviewed By: bmsdave

Differential Revision: D75606511

fbshipit-source-id: 9573f47d4622b7c8c1bc7cd51e9a15d5a7872aaf