Remove is_connect reference from old Hyper
Summary:
This doesn't appear to really be used (see the "allow unused" note in the
source - https://docs.rs/hyper/0.14.26/src/hyper/error.rs.html#38) and there is
no obvious analog for this value in the modern Hyper. We can drop this check
here separate of the upcoming Hyper changes.

Reviewed By: JakobDegen

Differential Revision: D75983586

fbshipit-source-id: b85a8c5175968cde7f2314b4994ef75b4a841c50