Unify bitcode optimization by defining artifacts for each extract bitcode file upfront
Summary:
Distributed thin-lto must create actions to optimize and codegen every bitcode file provided as input to the link. If pre built static archives are included in a link, information about the contents of these archives (ie. the name and number of archive members) is not stored in the graph. This information can only be determined via dynamic_output.

Until now, the approach to handle pre build static archives, has been to treat their contents in Starlark logic as a largely opaque directory of artifacts. For each pre built archive, we declare this directory as an artifact, and extract the contents of the archive into this directory. We also produce a manifest JSON file that describes the contents of this directory, and can be read via dynamic_output.

This means our starlark logic has no in memory representation for any given archive member, only this opaque directory of extracted archive members. Any actions created that need to operate on a single archive member, take the entire directory as input, and a command line flag somehow specifying what archive member to operate on.

This is bad for a number of reasons:

1. It unnecessarily bloats action inputs (you need to consume the entire directory to act on just one member).
2. It unnecessarily complicates actions. Each action needs to publish a manifest describing its outputs, and consuming actions need to read this manifest.
3. It unnecessarily makes downstream actions require dynamic_ouput (to read the manifest to act on directory contents). For example, this change makes thin-link a regular action, when it used to require dynamic_output.
4. It creates a lot of duplicated logic for operating on "standalone object files" and "extracted object files".

This problem is getting worse as we try to add support for double codegen, as we'll be chaining more of the actions that consume and produce opaque directories with manifests together.

As before we unarchive every archive and produce a manifest describing the contents. With this change, we introduce another class of `thin_lto_map_archive` action which consumes a directory, and maps its contents to individual artifacts declared in starlark.

We introduce `complete_distributed_link_with_expanded_archive_link_data` which is a function called via `dynamic_output`. It runs the aforementioned `thin_lto_map_archive` to declare and bind all the archive members. Since the artifacts we declare here cannot escape this function, the entire link is performed from here.

From this point forward, all the logic is agnostic to if an object file artifacts was generated by extraction from an archive, or not.

Reviewed By: rmaz

Differential Revision: D75806415

fbshipit-source-id: cc7f2725f59fa01c4b4f51e56e1a19d94292a139