Move native link flags into the argsfile
Summary: We have somewhat arbitrarily distributed some flags for native link in a response file, while leaving others on the command line. We currently place `common_link_flags` in the command directly rather than in the argsfile. In a future change, we want to include more flags in `common_link_flags`. To avoid making the command too verbose when that happens, include as much as possible in the argsfile.

Reviewed By: rmaz

Differential Revision: D75806409

fbshipit-source-id: 8243e1890ca1916a99512947b4fb6a6804a3281b