Map absolute paths from build script execution to rust compilation
Summary:
In general, the action that performs Cargo build script execution and the action that performs the corresponding Rust library compilation can run on different machines, with the repository checkout located at a different absolute path.

This means that when a Cargo build script tries setting environment variables which are absolute paths, as done by the `mime_guess2` crate here, they need to be rewritten to account for the new repository location before handing off to rustc.

https://www.internalfb.com/code/fbsource/[f8614eeba7dd3b9b32c267aef9937c1f457900b2]/third-party/rust/vendor/mime_guess2-2.0.5/build.rs?lines=24-25%2C28-31

This diff exposes a `$(abspath ...)` syntax that the Cargo build script runner will be able to use for such environment variables in {D76184973}.

Reviewed By: diliop

Differential Revision: D76184972

fbshipit-source-id: c99bbeb3c8575299431f7c0e0ab5220c254346ed