Set up conversion of new certificate/key errors
Summary:
The new versions of these packages have more specific error types which we
will need to accomodate. Note that this diff brings in the new versions of
these packages (BEFORE any other part of the project depends on them). This
should be fine because we're not doing changing something fundamental like
prost which would cause conflicts and it makes diff covering the actual
breaking changes more intelligible.

Reviewed By: Will-MingLun-Li

Differential Revision: D75961010

fbshipit-source-id: bb27e6f3476ec80ea19083ca12a50f429594465e