Add parsing modifiers logic to lex_provider_pattern
Summary:
Added the logic to parse `?modifiers` pattern to `lex_provider_pattern`. The decision has been made to assume that there will only be at most one `?` in the pattern, otherwise an error will be thrown.

Non-empty modifiers are incompatible with `PatternData<ConfiguredProvidersPatternExtra>` when `ConfigurationPredicate` is not `Any` or `Builtin` therefore we throw an error when this is the case in `lex_configured_providers_pattern`.

Reviewed By: christolliday

Differential Revision: D75878455

fbshipit-source-id: f902364d1cbe9cb589dd6282c9bc48426ff7d4d4