Add buckconfig for logging total configured graph unconfigured sketch
Summary:
We want to get the sketch of unconfigured targets reachable from a build via set sketch to compare against configured graph sketch so we can know whether a growth in configured graph size is coming from increase in configurations or increase in unconfigured targets. The following couple of diffs will all be adding this logging into build report. Buck2 dev post to follow shortly.

I thought it would be inaccurate to title this sketch unconfigured graph sketch since it technically does not traverse the unconfigured graph in the way that Buck typically does for commands like uquery. Instead, it traverses the configured graph but just sketches the unconfigured target labels. For that reason, I'm calling this sketch "configured graph unconfigured sketch", which is more accurate but rather verbose. Open to better name options here.

For now, we are just going to compute and log this sketch for all targets in the build command. This diff adds the buckconfigs to log this sketch in `-c buck2.log_total_configured_graph_unconfigured_sketch`.

Reviewed By: IanChilds

Differential Revision: D75423240

fbshipit-source-id: 146a8acc10f3cca6cd2ae71b34391e497319f1c0