Use experimental content-based path hashing in js_library
Summary:
I think this is a good candidate for turning content-based path hashing on for - the rules are mostly self-contained, do not use dep-files and they generally are used as inputs to other `js_library` rules, so we should see some cache hits across configurations, which is what we want.

I didn't do this for `js_bundle`, because it is used with `ensure_artifacts` in BXL, which is not yet supported.

Read more about this project here: https://fb.prod.workplace.com/groups/2567537826781973/posts/2641180899417665/

Reviewed By: blackm00n

Differential Revision: D76133687

fbshipit-source-id: b6d0812cd66f07099c77eabe2bdbf83fc6f3185a