Update parse_target_pattern to return ParsedPatternWithModifiers
Summary: Update `parse_target_pattern` to return `ParsedPatternWithModifiers` instead of `ParsedPattern`. This also involved integrating `ParsedPatternWithModifiers` into the existing parsing functions of `ParsedPattern` to handle the new struct. This was done by creating a new function for `ParsedPattern` which converts a `ParsedPatternWithModifiers` object to a `ParsedPattern` object if the modifiers is `None` otherwise we throw an error.

Reviewed By: Will-MingLun-Li

Differential Revision: D75919541

fbshipit-source-id: 22c81c4a1872cf26b5d0ef265ba8d0384300bae3