completions: Disable sorting in zsh
Summary:
A couple tests are failing because zsh completions are being emitted in unexpected orders. This diff disables automatic sorting for the completions, and instead just relies on the order that buck outputs. Even if that order isn't ideal, it'll be consistent between shells and we can fix it just once in buck

It's not clear to me why these tests only started failing recently and why this is not a problem on Linux. My guess is that the answer to both is that this is a new behavior in some recent zsh version

Reviewed By: IanChilds

Differential Revision: D76405387

fbshipit-source-id: 7d0281086d1f91e1eec35f94608fef374bd1d8bc