Use cache buster in test to avoid get a remote-dep-file hit from a different test
Summary:
Well, these tests sure are a pain, given you can't really test them properly locally and have to just infer what is going to happen once the upload happens on Sandcastle.

Anyway, the issue is that we are getting a dep-file hit from the final upload in `test_re_dep_file_query_change_tagged_used_file`, which then fails some checks. We can use a cache buster for just this test to make sure that does not happen.

Reviewed By: blackm00n

Differential Revision: D76417553

fbshipit-source-id: 5e854fb2d7ef1d30f52df7c17245bf58d9da5ad0