treat failed analysis and bxl computations as transient
Summary:
Cancellations of starlark analysis appear to be sticky. If an analysis fails, other analyses would be cancelled, and subsequent runs (that don't touch the inputs of the previously cancelled keys) would fail with cancellation errors.

Analyses that get cancelled are _supposed_ to not populate Dice, but it appears that this can (currently) be mis-detected.

Ideally, we'd treat _only_ cancellation errors as transient. However we don't expose that the underlying error is a cancellation error by the time we get to where we need to make this decision.

In the interest of not caching incorrect things, for now just treat any Bxl or Analysis error as transient, until we can fix it the right way.

As reported here: https://fb.workplace.com/groups/buck2users/permalink/28685212781100631/

Reviewed By: JakobDegen

Differential Revision: D76479520

fbshipit-source-id: 1cb4213013bde53ec8497f0360c50026d83084f3