S530607: [buck2] disable starlark cancellation
Summary:
Starlark cancellation bypassed the normal async-cancellation-processing, which had the unintended consequence of also bypassing the protection against cancelling tasks still needed by other computations.

Imagine you build A and B, A depends on C, and both A,B depend on D. B has an error, so it causes A to be cancelled. A no longer needs C, so it attempts to cancel it. Normally, C would determine it's still needed by B, and would not cancel. But starlark evaluation cancellation bypasses this check, and WILL cancel C, which in turn will induce an error on B that wouldn't have existed before.

Reviewed By: JakobDegen

Differential Revision: D76523417

fbshipit-source-id: ac789408cb081575186f4f5c4692f0a2645faa32