Implement rudimentary double codegen support
Summary:
This adds support for an unoptimized form of double codegen to the Darwin distributed thin-lto rules. Unlike local thin-lto which  dumps optimized bitcode, collects cgdata from codegening each module, then re-codegens the previously saved optimized bitcode this time with the cgdata, this implementation is a bit simpler.

For now, for each input bitcode file we compile it to a native object file. Each such produced native object file is fed into `llvm-cgdata --merge` to merge the cgdata stored in each object file. We then repeat the compilation of every input bitcode file, starting from unoptimized bitcode and re-compile it, this time with cgdata.

This is not optimal, ideally we should mirror the local thin-lto implementation, but that requires more toolchain changes. We do not at the moment have the necessary support in the toolchain to dump optimized bitcode truly right before codegen begins, no do we have the ability to instruct Clang to skip opt and just go straight to codegen.

This will unblock us for now, reducing reliance on asyncrhonously collected cgdata while we add toolchain support for a more optimized system.

Reviewed By: rmaz

Differential Revision: D75835088

fbshipit-source-id: 6069b0ca239d5ce7956e60720338b399da70a8e9