Fix up broken buck2_explain test
Summary:
Slightly ironic that buck2 explain test case started failing right after iguridi left.

The issue seems to be that BUCK_WRAPPER_UUID was not set for the first build command which launched the daemon, so env var was not propagated properly. Not sure why it was working before. Maybe someone was trying this locally and it caused the corresponding log to be populated in manifold and the TTL only expired recently.

Reviewed By: JakobDegen

Differential Revision: D76648150

fbshipit-source-id: ca556341358bbb1714e7e13db40e4fb1a2814086