Use ExitCode::Success if test runner returns exit code 0
Summary:
`ExitCode::TestRunner` is used to represent that exit codes from an external source (TPX) may not have the same semantics as ones generated from buck, but in the case of exit code 0 the meaning is the same.

More importantly, `ExitCode::Success` is used to determine if a command was successful when calculating `outcome` and `succeeded` so `TestRunner(0)` is broken. This is still a bit of a footgun so I'm planning to clean this up further in a follow up.

Reviewed By: JakobDegen

Differential Revision: D76919557

fbshipit-source-id: 2cd72d1167b52f54cf8b095beb6fd90540728e49