Better Error Tags for EventLog Errors
Summary: EVENT_LOG error types spiked in the past week https://fburl.com/unidash/i7icwcfs. It's kind of difficult to figure out which error type spiked exactly as there are 3 possible causes. Looking at scuba, it seems like it's `RecentIndexOutOfBounds` which is an Input error, tagging it as so

Reviewed By: rajneesh

Differential Revision: D76932437

fbshipit-source-id: adbebb14bb104e866031b1e7bdc6f01ad5d3d4f4