Split up InvocationRecorder initialization
Summary:
The goal of this stack is to move invocation record logging to the top of the stack in order to log errors in early start up so that we have error data for all failed commands, unless logging itself fails.

Doing this requires creating a mostly empty InvocationRecorder initially and updating it separately with command specific fields via `update_for_client_ctx`. For now just creating the most empty InvocationRecorder in the same place, later diffs cause it to be created earlier and logged later.

For now if `update_for_client_ctx` is not reached the command name will be 'unknown' and the cli args will be empty.

Reviewed By: rajneesh

Differential Revision: D75581279

fbshipit-source-id: 2ccd71dc267729bf61f0d784c6198e80a4e9cbd1