Merge EventSubscribers and EventsCtx
Summary:
`EventSubscribers` was added to handle client side events in contexts with no daemon connection, when `EventsCtx` was tied to a daemon connection. `EventsCtx` should now be tied to the lifetime of a client command so this separation doesn't make sense anymore.

Combining them to simplify refactoring in following diffs.

Reviewed By: JakobDegen

Differential Revision: D75581274

fbshipit-source-id: 4dfc9b09d65274ff7fa655c4fd36084751638483