add recommendations to TargetInfo
Summary:
This is probably one of the more controversial changes in this stack. I could see arguments for better naming or struct design.

Here, I'm plumbing a way for the BXL `TargetInfo` query to communicate a different build mode and/or build configuration flag recommendations (i.e. `-c debug.focused_folders=a,b,c`) to FDB.

I decided for a `struct` that wraps the existing `TargetInfo` instead of extending the `TargetInfo` since our `FbsourceExecInfo` contains `TargetInfo` as well. `FbsourceExecInfo` probably doesn't need to know about recommendations at that stage. I wouldn't be surprised if there was a better way to express this.

In this diff, the BXL just returns the "empty" values for `recommended_build_mode` and `recommended_build_config`. See the later diffs for how this is used.

Reviewed By: patskovn

Differential Revision: D76275177

fbshipit-source-id: 8724fb9af7329501ed88eeebc1cbcd662b4f4c76