Change delimiter for ?modifier syntax to +
Summary:
Change delimiter for ?modifier syntax to + as well as update tests to reflect this change.

This change was made to support the use of `?modifier` syntax with the `--target-universe` flag, which also uses `,` as a delimiter between universes therefore a different symbol must be used. A post discussing this in more detail can be found here:

https://fb.workplace.com/groups/buck2dev/permalink/4026976687590373/

Reviewed By: Will-MingLun-Li

Differential Revision: D76987667

fbshipit-source-id: d27daf8eea5b9836a2016cd6b106b136508cb3f6