Remove unnecessary and incorrect check for ? in cquery
Summary:
The check for `?` directly in the `cquery` command is unnecessary and seems to be causing issues with the `attrregexfilter` as seen here:

https://fb.workplace.com/groups/buck2users/permalink/4004624319793919/

Removed the check for `?` directly in the `cquery` command and let it error out during actual parsing.

Updated old test and added new test to ensure that the issue does not occur again.

Reviewed By: scottcao

Differential Revision: D77234653

fbshipit-source-id: fc6a5bc80a6b02e69ddb18ff564a42942ba076b9