Support selects when parsing parse_codesign_entitlements
Summary:
In apple_bundles, prior to defining the target, we inspect the info_plist_substitutions to capture any potential entitlements files we need as part of the target.

However, because we inspect the value, treating it as a string or dict, we can't handle selects that may be provided for variable values.

Reviewed By: milend

Differential Revision: D77317993

fbshipit-source-id: f7b758b8c610832eee009160c4623e98c1060730