artifacts: Introduce `ValueAsArtifactAnnotation`
Summary:
This is a marker type for `UnpackValue` impls for artifact-likes without a lifetime attached to it.

The reason that this is needed is because of existing callsites like this:

https://www.internalfb.com/code/fbsource/[f23289d5bad7985c5762b8476486d16dae79bf04]/fbcode/buck2/app/buck2_build_api/src/interpreter/rule_defs/provider/builtin/default_info.rs?lines=143

Note the `'static` lifetime - that lifetime is completely wrong. We happen to get away with it right now because of `.cast()` calls on `ValueOfUncheckedGeneric` and 2) the covariance of that parameter. However, the covariance is basically an accident, most `'v`s are invariant, and in a future diff that becomes true for these `'v`s as well.

Ideally, we would update this code to actually be able to name `'v` there - however, the `internal_provider` macro makes that extremely hard and would largely have to be re-written. Given that I am planning on enabling a bunch of simplifications to it anyway though, it doesn't make sense to do that right now, so introduce this instead as a workaround.

Differential Revision: D77119462

fbshipit-source-id: b633740bdf9b3cce350299058f0685e1ab954756