tests: Delete fat platform tests
Summary:
This set of tests "tests fat platforms" - but it has a couple of problems

First of all, it's really more of a demonstration than a test. The only code that it actually shares with the prod fat platform implemenation is this thing, and nothing else:

https://www.internalfb.com/code/fbsource/[c1964234c611]/fbcode/buck2/platform/execution/fat_binary.bzl?lines=20

This is a lot of test code for testing a 50 line rule.

That means it also doesn't actually cover most of the fat platform implementation, which is anyway well covered by eg the apple tests.

Second, in order to make this test work it has to run buck in fbsource while overriding `-c build.execution_platforms`. That's completely unsupported though and frankly will inevitably break sooner or later.

So just don't have these tests

Reviewed By: blackm00n

Differential Revision: D77397671

fbshipit-source-id: 14f72e5ba6898da54c6bd7bfa3c7c7f5f9263b26