artifacts: Remove useless sort-of-type-annotation from output artifact
Summary: A couple diffs down the stack`StarlarkDeclaredArtifact` gets a lifetime - that would result in us running into the same problem here that we ran into in the previous diff. So avoid that by just not adding the `ValueOfUncheckedGeneric`. Looking at the callsites reveals that this wasn't adding any type safety anyway

Reviewed By: cjhopman

Differential Revision: D77119458

fbshipit-source-id: a5a67a70a4dce6e6159ac35d02f286ce6a977cf2