artifacts: Attach a `'v` lifetime to `DeclaredArtifact`
Summary:
Declared artifacts have shared mutability during an analysis evaluation. This isn't all that uncommon, but what is uncommon is that they achieve that by just storing an `Rc`:

https://www.internalfb.com/code/fbsource/[72d70f64f6a9b1869a176e70a2ad2d6d95eb5729]/fbcode/buck2/app/buck2_artifact/src/artifact/artifact_type.rs?lines=341

We're not normally in the business of doing that. Shared mutability in starlark is generally represented by having some starlark value own a `RefCell`, allocating that starlark value, and then using references to that allocation to achieve sharing.

That approach is generally a bit better because it cleanly ties the "duration" of the shared mutability to the starlark evaluation - once the starlark evaluation has completed, the shared mutable thing is gone - this helps with correctness a bit (important, given the standard implications of shared mutability)

More concretely though, we have this thing:

https://www.internalfb.com/code/fbsource/[72d70f64f6a9]/fbcode/buck2/app/buck2_core/src/unsafe_send_future.rs?lines=38-47

Used here, for example:

https://www.internalfb.com/code/fbsource/[72d70f64f6a9]/fbcode/buck2/app/buck2_analysis/src/analysis/env.rs?lines=233

This justifying for Sendness is incorrect in cases like this. Even if you send an entire heap with all of its values across a thread, there's nothing guaranteeing that the `Rc` hasn't been cloned into something that isn't being sent across threads.

A future diff does a better job of using the type system (instead of unsafe) to justify `Send` in a situation like that, and it turned up this problem.

So the plan is to replace the `Rc` with an allocation that's correctly tied to the starlark heap. That requires putting a `'v` lifetime onto the `DeclaredArtifact` type, which in turn requires putting it onto a great many other places. This diff does that without yet using the `'v` lifetime for anything interesting, just a `PhantomData` for now.

All changes in this diff are trivial(ish). The interesting stuff was pulled out earlier in the stack

Differential Revision: D77119488

fbshipit-source-id: dbefb6b28a13fa166e0e88e4bda5fba73b329460