artifacts: Split declared and frozen artifacts in artifacts visitor
Summary:
When visiting a frozen output artifact in an artifacts visitor, we currently allocate a new declared artifact just for that purpose and return it

https://www.internalfb.com/code/fbsource/[72d70f64f6a9b1869a176e70a2ad2d6d95eb5729]/fbcode/buck2/app/buck2_artifact/src/artifact/artifact_type.rs?lines=302-308

This is pretty misleading, as the `OutputArtifact` type does not typically exist for artifacts that are not being bound in the current analysis.

So instead, have two methods on the trait to distinguish them properly.

This reveals that `cmd_args.output` allows you to get an `OutputArtifact` for an artifact defined in a dependent analysis in starlark, which seems bad.

Reviewed By: IanChilds

Differential Revision: D77119485

fbshipit-source-id: d8396ba6341dd6d4d171027cbb3cc58401c92c8a