Mockingbird: Allow Parallel Mock Generation
Summary: This line I'm deleting made it so mockingbird would only generate one mock at a time. It was added because we'd get occasional instances where mockingbird would OOM and get stuck. A lot has changed in the framework since then so I think we might be alright without this now. It being there means we're not building as fast as we could be. Removing it and will monitor to see if problems arrise.

Reviewed By: milend

Differential Revision: D77594640

fbshipit-source-id: a9688b62eabecf0ccd4f443022bff2a705e624d6