Split protobuf generation rule to allow modern prost.
Summary: This sets us up to introduce another definition using the modern prost (and/or dynamic dep surgery) in a followup. The overall strategy is still being worked out but we will need to do (at least) one of the above (since the alternative is to migrate the entire repo to the new prost in one diff) so we might as well lay the groundwork.

Reviewed By: JakobDegen

Differential Revision: D75883484

fbshipit-source-id: 12b9aabfa24204c0c6a76f4ab50d6e64cee51305