add build mode and focus recommendations to target_info.bxl
Summary:
In this diff, `target_info.bxl` now starts reading the label for the recommended build mode, and it looks for the `fdb:focus_auto=true` label for first-level dependencies.

In `focus_auto.bzl`, I included the logic of getting the first level dependencies for a target. We perform some adhoc filtering to remove dependencies like from `fbcode_macros//` that are present in a lot of targets but are unlikely what users want to debug (and thus good candidates for optimizing).

Reviewed By: patskovn

Differential Revision: D76176676

fbshipit-source-id: 592c0ea41b5927516d725d2e4823d199c9c4e6d2