allow_cache_upload if using late_build_info_stamping
Summary: Similar to the previous diff, we should allow_cache_upload on CI if a cxx_binary or cxx_unittest enables late build info stamping; this allows users to skip relinking if they are on stable with no local changes. This diff focuses on the distributed thinlto use case, particularly helpful for services that need hotfixes with lto.

Reviewed By: AishwaryaSivaraman

Differential Revision: D77674363

fbshipit-source-id: 4b8bd298d475197f22af5a3eb06b898f184237c7