Disable wrongly passing test
Summary: Test was originally set up to expect failure because the expected exit code wasn't available in Eden. It has since been added so the build should have started to succeed (And therefore causing the test to fail). This didn't happen and Eden was still failing with the old exit code. Reached out to Eden to figure out why but disabling it for now

Reviewed By: IanChilds

Differential Revision: D77756099

fbshipit-source-id: 9624fb618614470d0812e02cb1a5494826a5a7ba