Add repro test case for bug in new_platform_hash_rollout
Summary:
`buck2.new_platform_hash_rollout` buckconfig is used to turn on strong hashing for buck-out hashes, which would eliminate the issue of hash collisions. Unfortunately, I realize there is a bug in which the rollout would cause a graph divergence problem - Because `new_platform_hash_rollout` is only read by core but only on daemon startup, a buck daemon used after landing changes to `buck2.new_platform_hash_rollout` could still be using previous weak hashes, but if this daemon gets killed and restarted, then it would use strong hashes on the same revision, diverging the build graph on the same revision and creating a build speed/capacity problem.

This diff adds a test case to repro this bug.

Reviewed By: JakobDegen

Differential Revision: D77637075

fbshipit-source-id: 82699b911863d13d73fbd0abdd737c4e988351d6