Fix bug in new_platform_hash_rollout
Summary:
For more context on what this bug is, see description in D77637075.

Add `new_platform_hash_rollout` to `DaemonStartupConfig`, such that if `new_platform_hash_rollout` buckconfig changes it forces a daemon restart. This will ensure that daemon is always using the correct hash based on value of buckconfig.

Note this value isn't actually read from `DaemonStartupConfig`. We just need it to force a daemon restart, and the logic to wire this together is very convoluted. This will all be deleted once rollout for new hash finishes.

Reviewed By: JakobDegen

Differential Revision: D77637458

fbshipit-source-id: cc027f53f5cba5b270c2b01b47633189ca48399e