Map content based paths to constant value when running dep-files
Summary:
Problem: if we have a dir with a content-based path and two entries `dir/tagged.txt` and `dir/untagged.txt`, then if we change `untagged.txt`, the path to `dir/tagged.txt` will also change, and when we try to partition the inputs using the previous dep-file, we won't find the path at all. Also, when we "fingerprint" the tagged path, it will give a different value since the path is part of the fingerprint.

Right now we store the partitioned inputs using the content-based paths. Instead of that, we store them using a dep-files placeholder instead of any content-based paths (that's basically what `add_to_directory_for_dep_files` is doing here).

Of course, actions write dep-files using the content-based paths that they are given, so when we read the dep-files, we need to translate them to use the dep-files placeholder. We do that using string matching which finds the content-based hash and changes it to a placeholder. If that string matching doesn't get what we expect, then we just soft error and throw away the dep-files.

Reviewed By: cjhopman

Differential Revision: D75526488

fbshipit-source-id: 5cfbe0100e197f7ed6b9bf4899ddc3c00f51f824